backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
google-auth==2.32.0
requests==2.32.3
starlette==0.37.2
snowflake-connector-python==3.17.2
google-genai
numpy
opencv-python-headless
